moves shared to video folder
