+++ modules/videoModule/shared/parentModule.js
