+++ modules/videoModule/shared/state.js
